app: runCmd: Extract handleRunCommand functionality from main.ts
Move functionality for this feature into its own module.
