+++ app/.gitignore
@@ -8,5 +8,6 @@
 /electron/preload.js
 /electron/windowSize.js
 /electron/plugin-management.js
+/electron/runCmd.js
 electron/windowSize.test.js
 electron/env-paths.js
