+++ app/electron/main.ts
@@ -25,6 +25,7 @@ import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 import i18n from './i18next.config';
 import { PluginManager } from './plugin-management';
+import { handleRunCommand } from './runCmd';
 import windowSize from './windowSize';
 
 dotenv.config({ path: path.join(process.resourcesPath, '.env') });
@@ -1201,144 +1202,7 @@ function startElecron() {
       mainWindow?.webContents.send('backend-token', backendToken);
     });
 
-    /**
-     * Data sent from the renderer process when a 'run-command' event is emitted.
-     */
-    interface CommandData {
-      /** The unique ID of the command. */
-      id: string;
-      /** The command to run. */
-      command: string;
-      /** The arguments to pass to the command. */
-      args: string[];
-      /**
-       * Options to pass to the command.
-       * See https://nodejs.org/api/child_process.html#child_process_child_process_spawn_command_args_options
-       */
-      options: {};
-    }
-
-    /**
-     * Ask the user with an electron dialog if they want to allow the command
-     * to be executed.
-     * @param command - The command to show in the dialog.
-     * @returns true if the user allows the command to be executed, false otherwise.
-     */
-    function confirmCommandDialog(command: string): boolean {
-      if (mainWindow === null) {
-        return false;
-      }
-      const resp = dialog.showMessageBoxSync(mainWindow, {
-        title: i18n.t('Consent to command being run'),
-        message: i18n.t('Allow this local command to be executed? Your choice will be saved.'),
-        detail: command,
-        type: 'question',
-        buttons: [i18n.t('Allow'), i18n.t('Deny')],
-      });
-
-      return resp === 0;
-    }
-
-    const SETTINGS_PATH = path.join(app.getPath('userData'), 'settings.json');
-
-    /**
-     * Loads the user settings.
-     * If the settings file does not exist, an empty object is returned.
-     * @returns The settings object.
-     */
-    function loadSettings() {
-      try {
-        const data = fs.readFileSync(SETTINGS_PATH, 'utf-8');
-        return JSON.parse(data);
-      } catch (error) {
-        return {};
-      }
-    }
-
-    /**
-     * Saves the user settings.
-     * @param settings - The settings object to save.
-     */
-    function saveSettings(settings) {
-      fs.writeFileSync(SETTINGS_PATH, JSON.stringify(settings), 'utf-8');
-    }
-
-    /**
-     * Checks if the user has already consented to running the command.
-     *
-     * If the user has not consented, a dialog is shown to ask for consent.
-     *
-     * @param command - The command to check.
-     * @returns true if the user has consented to running the command, false otherwise.
-     */
-    function checkCommandConsent(command: string): boolean {
-      const settings = loadSettings();
-      const confirmedCommands = settings?.confirmedCommands;
-      const savedCommand: boolean | undefined = confirmedCommands
-        ? confirmedCommands[command]
-        : undefined;
-
-      if (savedCommand === false) {
-        console.error(`Invalid command: ${command}, command not allowed by users choice`);
-        return false;
-      } else if (savedCommand === undefined) {
-        const commandChoice = confirmCommandDialog(command);
-        if (settings?.confirmedCommands === undefined) {
-          settings.confirmedCommands = {};
-        }
-        settings.confirmedCommands[command] = commandChoice;
-        saveSettings(settings);
-      }
-      return true;
-    }
-
-    /**
-     * Handles 'run-command' events from the renderer process.
-     *
-     * Spawns the requested command and sends 'command-stdout',
-     * 'command-stderr', and 'command-exit' events back to the renderer
-     * process with the command's output and exit code.
-     *
-     * @param event - The event object.
-     * @param eventData - The data sent from the renderer process.
-     */
-    function handleRunCommand(event: IpcMainEvent, eventData: CommandData): void {
-      return; // Disable this until we figure out a better way to do this
-
-      // Only allow "minikube", and "az" commands
-      const validCommands = ['minikube', 'az'];
-
-      if (!validCommands.includes(eventData.command)) {
-        console.error(
-          `Invalid command: ${eventData.command}, only valid commands are: ${JSON.stringify(
-            validCommands
-          )}`
-        );
-        return;
-      }
-      if (!checkCommandConsent(eventData.command)) {
-        return;
-      }
-
-      const child: ChildProcessWithoutNullStreams = spawn(eventData.command, eventData.args, {
-        ...eventData.options,
-        shell: false,
-      });
-
-      child.stdout.on('data', (data: string | Buffer) => {
-        event.sender.send('command-stdout', eventData.id, data.toString());
-      });
-
-      child.stderr.on('data', (data: string | Buffer) => {
-        event.sender.send('command-stderr', eventData.id, data.toString());
-      });
-
-      child.on('exit', (code: number | null) => {
-        event.sender.send('command-exit', eventData.id, code);
-      });
-    }
-
-    ipcMain.on('run-command', handleRunCommand);
+    ipcMain.on('run-command', (event, eventData) => handleRunCommand(event, eventData, mainWindow));
 
     new PluginManagerEventListeners().setupEventHandlers();
