+++ app/package.json
@@ -120,7 +120,8 @@
       "electron/i18n-helper.js",
       "electron/windowSize.js",
       "electron/env-paths.js",
-      "electron/plugin-management.js"
+      "electron/plugin-management.js",
+      "electron/runCmd.js"
     ],
     "extraResources": [
       {
